Questionaire simple paginator
